파일 간의 패키지 설정
