--- backend/src/main/java/com/woochacha/backend/domain/member/entity/BranchList.java
@@ -1,5 +1,0 @@
-package com.woochacha.backend.domain.member.entity;
-
-public enum BranchList {
-    서울, 대전, 부산
-}
