+++ backend/src/main/java/com/woochacha/backend/domain/member/entity/branch/BranchList.java
@@ -1,0 +1,5 @@
+package com.woochacha.backend.domain.member.entity.branch;
+
+public enum BranchList {
+    서울, 대전, 부산
+}
