+++ backend/src/main/java/com/woochacha/backend/domain/member/entity/carentity/FuelList.java
@@ -1,4 +1,4 @@
-package com.woochacha.backend.domain.member.entity.CarEntity;
+package com.woochacha.backend.domain.member.entity.carentity;
 
 public enum FuelList {
     가솔린, 디젤, LPG, 가솔린ANDLPG, 하이브리드WITHLPG, 하이브리드WITH가솔린, 하이브리드WITH디젤, CNG, 전기, 기타
