+++ backend/src/main/java/com/woochacha/backend/domain/member/entity/carentity/Transmission.java
@@ -1,4 +1,4 @@
-package com.woochacha.backend.domain.member.entity.CarEntity;
+package com.woochacha.backend.domain.member.entity.carentity;
 
 import lombok.Getter;
