+++ backend/src/main/java/com/woochacha/backend/domain/member/entity/carentity/TransmissionList.java
@@ -1,4 +1,4 @@
-package com.woochacha.backend.domain.member.entity.CarEntity;
+package com.woochacha.backend.domain.member.entity.carentity;
 
 public enum TransmissionList {
     오토, 수동, 세미오토, CVT, 기타
