+++ backend/src/main/java/com/woochacha/backend/domain/member/entity/carentity/TypeList.java
@@ -1,4 +1,4 @@
-package com.woochacha.backend.domain.member.entity.CarEntity;
+package com.woochacha.backend.domain.member.entity.carentity;
 
 public enum TypeList {
     경차, 소형, 준중형, 중형, 대형, 스포츠카, RV, SUV, 승합, 버스, 트럭, 특수, 중기
