+++ backend/src/main/java/com/woochacha/backend/domain/member/entity/cartrade/Transaction.java
@@ -1,4 +1,4 @@
-package com.woochacha.backend.domain.member.entity;
+package com.woochacha.backend.domain.member.entity.cartrade;
 
 import lombok.Getter;
 import lombok.Setter;
